commit: Created simple navigation
